Use the same DialOptions for WASM and non-WASM
