+++ dial.go
@@ -20,36 +20,6 @@ import (
 	"github.com/coder/websocket/internal/errd"
 )
 
-// DialOptions represents Dial's options.
-type DialOptions struct {
-	// HTTPClient is used for the connection.
-	// Its Transport must return writable bodies for WebSocket handshakes.
-	// http.Transport does beginning with Go 1.12.
-	HTTPClient *http.Client
-
-	// HTTPHeader specifies the HTTP headers included in the handshake request.
-	HTTPHeader http.Header
-
-	// Host optionally overrides the Host HTTP header to send. If empty, the value
-	// of URL.Host will be used.
-	Host string
-
-	// Subprotocols lists the WebSocket subprotocols to negotiate with the server.
-	Subprotocols []string
-
-	// CompressionMode controls the compression mode.
-	// Defaults to CompressionDisabled.
-	//
-	// See docs on CompressionMode for details.
-	CompressionMode CompressionMode
-
-	// CompressionThreshold controls the minimum size of a message before compression is applied.
-	//
-	// Defaults to 512 bytes for CompressionNoContextTakeover and 128 bytes
-	// for CompressionContextTakeover.
-	CompressionThreshold int
-}
-
 func (opts *DialOptions) cloneWithDefaults(ctx context.Context) (context.Context, context.CancelFunc, *DialOptions) {
 	var cancel context.CancelFunc
